Validando remoção de itens de um pedido

+++ src/services/order/RemoveItemService.ts
@@ -1,16 +1,29 @@
 import { prisma } from '../../prisma';
 
-interface ItemRequest{
+interface RemoveItemRequest {
   item_id: string;
 }
 
 export class RemoveItemService {
-  async execute({ item_id }: ItemRequest) {
-    const order = await prisma.item.delete({
-      where:{
+  async execute({ item_id }: RemoveItemRequest) {
+    // Verifica se o item existe
+    const itemExists = await prisma.item.findFirst({
+      where: {
         id: item_id
       }
-    })
-    return order;
+    });
+
+    if (!itemExists) {
+      throw new Error('Item não encontrado');
+    }
+
+    // Remove o item
+    const item = await prisma.item.delete({
+      where: {
+        id: item_id
+      }
+    });
+
+    return item;
   }
 }
